+++ tgemu/src/types.h
@@ -4,11 +4,11 @@
 
 typedef unsigned char uint8;
 typedef unsigned short int uint16;
-typedef unsigned long int uint32;
+typedef unsigned int uint32;
 
 typedef signed char int8;
 typedef signed short int int16;
-typedef signed long int int32;
+typedef signed int int32;
 
 #endif /* _TYPES_H_ */
